Button Letter Spacing
The css class .spaced-text was applied to the parent container anchor tag, not the container bold tag. This caused the class to adjust the margin incorrectly for the letter spacing added to the last letter, instead affecting the entire button and not centering the text as desired. Class moved to bold tag.